Relocate RequestCoop controller + presenter
