+++ arbeitszeit_web/presenters/request_cooperation_presenter.py
@@ -1,17 +1,10 @@
 from dataclasses import dataclass
 from html import escape
-from typing import List, Protocol, Union
-from uuid import UUID
+from typing import List
 
-from arbeitszeit.use_cases.request_cooperation import (
-    RequestCooperationRequest,
-    RequestCooperationResponse,
-)
+from arbeitszeit.use_cases.request_cooperation import RequestCooperationResponse
 from arbeitszeit_web.email import EmailConfiguration, MailService
 from arbeitszeit_web.translator import Translator
-
-from .malformed_input_data import MalformedInputData
-from .session import Session
 
 
 @dataclass
@@ -101,43 +94,4 @@ class RequestCooperationPresenter:
             )
             % dict(coordinator=escape(coordinator_name)),
             sender=self.email_configuration.get_sender_address(),
-        )
-
-
-class RequestCooperationForm(Protocol):
-    def get_plan_id_string(self) -> str:
-        ...
-
-    def get_cooperation_id_string(self) -> str:
-        ...
-
-
-@dataclass
-class RequestCooperationController:
-    session: Session
-    translator: Translator
-
-    def import_form_data(
-        self, form: RequestCooperationForm
-    ) -> Union[RequestCooperationRequest, MalformedInputData, None]:
-        current_user = self.session.get_current_user()
-        if current_user is None:
-            return None
-        try:
-            plan_uuid = UUID(form.get_plan_id_string())
-        except (ValueError, TypeError):
-            return MalformedInputData(
-                "plan_id", self.translator.gettext("Invalid plan ID.")
-            )
-        try:
-            cooperation_uuid = UUID(form.get_cooperation_id_string())
-        except ValueError:
-            return MalformedInputData(
-                "cooperation_id",
-                self.translator.gettext("Invalid cooperation ID."),
-            )
-        return RequestCooperationRequest(
-            requester_id=current_user,
-            plan_id=plan_uuid,
-            cooperation_id=cooperation_uuid,
         )
