+++ tests/presenters/test_request_cooperation_presenter.py
@@ -2,7 +2,9 @@ from typing import Optional
 from unittest import TestCase
 
 from arbeitszeit.use_cases.request_cooperation import RequestCooperationResponse
-from arbeitszeit_web.request_cooperation import RequestCooperationPresenter
+from arbeitszeit_web.presenters.request_cooperation_presenter import (
+    RequestCooperationPresenter,
+)
 from tests.email import FakeEmailSender
 from tests.translator import FakeTranslator
